file wrappers: LZ4 Fast seeking of linked blocks
LZ4 1.10.0 promotes to the stable API a function to set the
dictionary. It only has an effect when decompression is in the
init state, but by resetting the decompression context and loading
our saved frame header, we can use this to fast seek in the same
way as gzip with zlib.

Saving with linked blocks has a bigger advantage with small blocks,
which can occur with live captures (because we need to flush when
packets are written, and packet throughput can be a lot lower than
a block size.) So we probably should switch to writing linked blocks
for live captures at least at some point, though it's tricky since
1.10.0 is so new.