JPEG: Workaround bogus EXIF IFDs generated by Flutter Image Picker
Flutter Image Picker generates non-conformant EXIF IFDs by omitting the
4-byte offset to the next IFD, which is usually zero. This can be
verified via the EXIF specification and the examples in Annex A.

We already check if the value has incorrectly gone backwards. Also
check if it is unreasonably too high, which catches at least some
(most?) cases.

Also update the URLs for the EXIF standards; https://www.exif.org
was an independent site that is no longer active. For the Wayback
Machine link, make sure to use a version before the site was taken
over by a domain squatter that left just enough of the original text
for SEO purposes but removed the actual specifications. Also link
to CIPA, which makes the specifications available for free. (JEITA
charges for them.)

Fix #20355