test: Fix skipping Rawshark tests on Big Endian
pytest.skip can't be used at the class scope, only inside a test.
At the class level the decorator should be used instead.
See the warning in the Debian s390x builds:
https://buildd.debian.org/status/fetch.php?pkg=wireshark&arch=s390x&ver=4.4.3-1&stamp=1736436117&raw=0

None of Rawshark is going to work on Big Endian, so change it to
a decorator for the class instead of moving it inside the test.

Ping #19965