PNIO: Read GSD XML files only when changing prefs
Only read the directory of GSDML files when prefs are changed,
loading into memory the structures that will be used. Use
libxml2 instead of manually parsing the files. (Yes, that means
we lose support for parsing the files if we don't have libxml2,
but we usually do, our official packages do, and we might want
to make it required, see issue #20166.)

This reduces the dissection time for a sample file from 12.8 seconds
to .020 seconds, because every file in the directory isn't parsed
anew for every single submodule in every single frame.

Use a wmem scope so that they are freed when the prefs are changed.

Make the field where the used GSD file is listed marked as generated,
and add the filename as a string instead of using item_append_text.
The error message field should probably be an expert info instead.

Check whether Wireshark is built with libxml2 support, and add some
messages if not.

Fix #19456