test: Skip tshark extcap test if sampleif.py execute bit is clear
Extcaps have to be executable, so sampleif.py has the execute bit
set in the git repository, which is also applied to clones - unless
the git config core.fileMode is set to false.

Since it's a Python extcap instead of a binary, we could run the
script anyway through an interpreter, but not for testing tshark
loading the interface end to end. Just skip the test.