tls-utils: Fix exporting secrets with renegotiation
Close the client random when inserting it into the hash map of
used client randoms. Just inserting the pointer means that when
renegotiation occurs, the data changes, which causes the last
used Client Random to be exported N times instead of the N different
Client Randoms.

This can be tested with the tls-renegotiation.pcap test capture
in the repository.

Ping #20362