DTLS: Decrypt DTLS 1.2 after renegotiation
DTLS, unlike TLS, only attempts decryption if SSL_HAVE_SESSION_KEY
is set (for 1.2). However, when renegotiation starts, the
SSL_HAVE_SESSION_KEY bit gets cleared from the state by
ssl_reset_session, indicating that the session key needs to be
re-derived.

At this point, the session is in a liminal state with respect
to the session key - the old session key needs to be used for
decrypting Encrypted Handshake Messages in order to get the new
key, but the bit needs to be clear in order for ssl_finalize_decryption
to derive the new session key.

TLS does not look at SSL_HAVE_SESSION_KEY before deciding whether
or not to attempt decryption. Have DTLS do the same thing.

DTLS has a special check for whether or not the cipher has null
encryption, while TLS doesn't but packet-tls-utils.c handles
ENC_NULL cipher suites appropriately. Remove the special DTLS
handling and behave as TLS does, so as to remove a test on
SSL_HAVE_SESSION_KEY.
(The DTLS special handling for NULL ciphers was added several years
earlier than the TLS generalized null cipher handling, compare commits
48d421ec9987409df46379e7ae0cf6c081491769 and
126c94d8c834538e3a2bc4bfe8f9534d6c189de5 and
8058ba646aee63d13405d8b7152a6aeb37fb6113
to
f96c56103b12302b2b235382627f8c035a890bdf
)

The NULL cipher DTLS file from #5863 still works with this change.

Fix #20362

Note: we likely still need the work for TLS in #18867 for handling a
case where a new cipher suite was selected in renegotiation than
chosen for the first handshake. That doesn't occur in the sample
for this issue.