macos-setup.sh: Set --build for x86_64 GMP build.
Without specifying a CPU family, the GMP configure script configures for the
CPU type of the build host.  This can break things for users with older Intel
CPUs when we have a builder with a newer CPU.

For x86_64, set "nehalem" as the CPU architecture, since that's the oldest CPU
architecture supported by the oldest version of macOS we presently support.
Since Apple is no longer using Intel CPUs, this can probably remain as-is until
we stop building for Intel Mac.

Tidy up a few comments.

Fixes #20348.