CMake: Require libpcap 1.5 or later
Require libpcap 1.5 (released 2013-11-07) or later. This rules
out building with the WinPCAP SDK.

This leaves the only optional libpcap features as remote capturing
support (via pcap_open) and the the sampling options (which are
treated as remote options and only enabled if we have remote
capture support).

Add a comment about how the workaround for building on macOS with
a SDK which has pcap_set_tstamp_precision and running on an earlier
version whose system shared libpcap doesn't support it shouldn't
be needed anymore, because we only support running on new enough
macOS.

Fix #20261