CMake: Adapt to minizip-ng 4.0.8 include file changes
At least since minizip-ng 2.10.2, minizip-ng, when built with the
original minizip compatibility layer, has had include files with
the original zip.h and unzip.h names from minizip.

Prior to 4.0.8, these all just included the entire mz_compat.h header;
in 4.0.8 the mz_compat.h header was removed in favor of having only
the matching compatible files, each of which now only includes the same
subset of functions and symbols as in the original minizip.

Since the zip.h and unzip.h files have long existed, just search for
and include those, since mz_compat.h has gone away in 4.0.8.

By default, when minizip-ng is built with the compatibility layer,
the library names itself minizip instead of minizip-ng so it can
be used as a drop in replacement. However, some packagers, e.g.,
Homebrew, build the compatibility layer but keep the -ng suffix
(or even perhaps the short-lived minizip2 name) so that minizip
and minizip-ng can co-exist on the same system.

We prefer a library called minizip-ng *with* the compatibility
layer to a library called minizip (which is presumably the old,
slower version) on the same system, but prefer a library called
minizip (which might be actually minizip-ng with the compatibility
layer) to a library called minizip-ng *without* the compatibility
layer.

That's for now, at least unless we port our code to use the minizip-ng
API.

We probably could combine these into one CMake find module, but it
might be easiest with the VALIDATOR option added to CMake 3.25 that
would let use reject minizip-ng without the compatibility layer.


(cherry picked from commit c3b410b50b2449519de899117a024007d9df29d3)

Co-authored-by: John Thacker <johnthacker@gmail.com>