Qt: Fix leak in Conversation and Endpoint recent tabs
g_list_prepend creates a new list link with the data provides,
so the old link has to be freed. Inserting the old link into the
list directly is more efficient, and newer GLib provides an API
call for that.

On older GLib we could manipulate the various list element prev
and next pointers directly, since they are exposed in the API,
to avoid creating a new link and freeing the old one, but this
is a rare operation so it's probably not worth making the code
harder to read.

Fixup 138d0438865d22cd1f09e2023fe2fd61bfe151ad