Use the Python UTF-8 mode for `asn2wrs.py`
Some of the input files to this script contain non-ASCII characters, encoded
with UTF-8. On Windows, the default encoding is usually not UTF-8, so reading
files opened without an explicit `encoding` argument can result in
`UnicodeDecodeError` exceptions. Enabling the UTF-8 mode makes UTF-8 the default
encoding regardless of the system encoding, preventing this problem.

The UTF-8 mode was added in Python 3.7. To keep compatibility with Python 3.6,
enable it via the `PYTHONUTF8` environment variable (rather than the `-X utf8`
option), since the variable will just be ignored by versions of Python
that don't support it.