Qt: Don't reset the model when only changing some data
When changing whether names are resolved, absolute time is
used, or the timestamp resolution in the Conversation or Endpoints
dialogs, don't reset the entire model, only emit a dataChanged
signal.

When new data arrives, call layoutChanged instead, which is a
little lighter than modelReset and will update all the indices.
The underlying GArray from from epan/conversation_table.c is
always the same object in a tap update.

Among other things, this keeps the selected and current rows in
the table the same instead of clearing them. For layoutChanged,
scroll so that the selected index (if there is one) stays visible.

Ping #16925 (this is an easier case than when applying a filter).