QUIC: Use connection ID, not ports and addresses, for reassembly
Implement custom reassembly functions that use the (internal) QUIC
connection ID and whether a frame is from the server or not, instead
of the current packet_info struct ports and addresses, for reassembly.

This should be more robust if STREAM frames are (ugh) fragmented across
migrated endpoints. This also should make reassembly unaffected by the
packet_info struct source or destination addresses or ports being
changed by an encapsulated protocol when there is more than one STREAM
frame in a single datagram.

Fix #20371