TCP: Initialize the per packet data to zero
Initialize the per packet data to zero when creating it. This avoids having
to remember to initialize new struct members when adding them to the struct.
(Such as the karn flag in the master branch, which apparently in some
circumstances with fuzzed data was used uninitialized.)

There are likely some other issues surrounding this per-packet data, since
its existence depends on the timestamp and packet number preference being set.

Also, change around the initialization so as to restore the stream packet
number on the second and later passes, which was accidentally removed
in 7ea8490ea86294f9990fb6ca59a7d2ff7764ee28.

Fix #20300