taps: Add limit to main filter flag
Add a tap listener flag for tap listeners whose results should
be limited automatically by the main display filter instead of
(or in addition to) their own filter. This allows such tap
listeners to update automatically to the new main display filter,
instead of waiting until after filtering completes (with themselves
using the old filter) and then retapping with the new filter.

Set it in the Conversations and Endpoints Dialogs when
"limit to display filter" is checked, so that changing the display
filter does not require an extra entire set of packet dissection
and retapping post display filter change per dialog with "limit to
display filter" checked. (Once to change the main filter, and then
one retap per dialog.) (See #20191) Instead, the Conversations
and Endpoints dialogs are retapped with the new filter at the same
time as the packet list.

The TrafficTab and ATapDataModel still have the functions to
set a per-tap display filter in addition to the main display filter,
but those are not used anymore. They could be in the future, perhaps.

In a future commit, this can be used towards #16925 and #18292
by only resetting and retapping tap listeners that depend on the
main filter (and dissector helpers) when the main display filter
changes but otherwise redissection is the same. (I.e., when
rescan_packets is called with redissect false, and therefore
add_packet_to_packet_list is called when add_to_packet_list false.)

Fix #20376