proto: add faster deregister for header fields
proto_deregister_field searches all fields of a parent to remove a
field. When removing a lot of dynamically registered fields, this is
very slow.

This patch adds a possibility to remove many fields in linear time.

This tremendously speeds up the removal of dynamically registered
fields with high numbers (e.g., 150k fields registered).

Signal PDU and SOME/IP are examples that profit from this change.