HTTP: Handle CONNECT method tunnelling the same port
Instead of just replacing the port with the tunnel port, also replace
the appropriate address with a string-based address from the hostname
in the CONNECT method response. This prevents the issue of the
decode_tcp_ports (and other functions) from finding the original
conversation when the proxied port is the same as the original port.
That prevents a possible infinite loop (#8733), finding the original
conversation dissector (HTTP instead of heuristically searching for TLS,
because many systems have an unencrypted CONNECT followed by switching
to TLS via proxy on port 443), or setting a conversation dissector in
the proxied dissector overriding the original HTTP dissector.

Tested with the file from #8733 as well.

Fix #20378.