RTP streams: Use limit to display filter flag for tap listener
Use the new limit to display filter flag for the tap listener when
the Limit to Display Filter box is checked, so that the tap
uses the main display filter at the same time as the main display
filter is applied to the packet list, instead of tapping occuring
with the old filter first, and then a tap with the new filter.

This removes an additional tap when the main display filter is changed.