HiPerConTracer: Disable UDP and TCP heurisics by default
Now that the HiPerConTracer dissector allows nanosecond timestamps
(commit 84df2b629ddf2a569f6d046d6e243f6fc279933b), the heuristic
is very weak. Previously, the 8 bytes that could contain the timestamp
had to be plausible as microseconds, specifically being between
1238799600000000 and 3889637999999999, or 0.01437% of the search space.
Now that nanoseconds are also allowed, values between 1238799600000000000
and 3889637999999999000 are allowed, which is 14.37% of the search
space.

That's too many false positives to enable by default on UDP and TCP.
Disable it until a more discriminating heuristic is available.