VOIP Calls: Use limit to display filter flag for tap listener
Use the new limit to display filter flag for VOIP Calls, so that
when the Limit to Display Filter box is checked, the tap uses
the main display filter when the main display filter is changed,
instead of using whether the old filter passed or not, and then
requiring another tap.

Move the call to the dialog's removeAllCalls() inside tapReset,
instead of requiring that anything that retaps have to remember to
call it.