CMake: Require 3.20 for cmake_path
Commit 9d8da5d95aab52da50c789d60a253ace17c43be8 uses cmake_path,
which was added in CMake 3.20. Make the requirement explicit instead
of failing in the middle of the script on older versions.

Among popular Linux distributions, the biggest impact is to Ubuntu
20.04 focal, which already requires unofficial packages to get a
supported Qt version and ends standard support in May 2025, and to
Debian Bullseye (current oldstable), where the CMake from backports
will be needed. Note that while OSS-Fuzz bases its containers on
Ubuntu 20.04 for now, a more recent CMake (currently 3.29.2) is
installed and used.

RHEL 8 and SLES 15 have had CMake 3.20 for several point releases
(8.5 and 15SP4), so it should not be a problem on updated systems.