Websocket: Fix decompression flag handling
The first fragment of a multiframe Websocket message is not
necessarily the first fragment of the Websocket conversation.
We need to save the opcode and per message compression flag from
the first fragment of the reassembled message, not the first
frame of the conversation. Some conversations may have control
frames.

Also, since according to RFC 6455 5.5, control frames can be
interjected in the middle of a fragmented message, can have
a payload, and MUST NOT be fragmented, we do not save the first
frag payload and per message compression flag until after checking
if the frame is a control frame.

Use proto_tree_add_ret_boolean in a couple places where possible.

Fixup aa2b43b7456539947fc7756bd3297ea3cbbde0bb

Fix #20390