Qt: Don't crash when trying to sort while sorting
Don't change the value of PacketListModel class private member variables
if we're not actually going to perform the sort. In particular, we want
to avoid the case where we're not sorting because we're in the middle of
a sort. The sort comparison function uses the class private member
variables to make the comparison. Changing them while the sort is
ongoing will make the sort invalid, and probably violate assumptions that
std::sort makes, leading to a crash.

We might be able to handle queuing another sort while a sort is on-going
if the comparison function were made a closure instead. (Though this
would still probably not be great for performance.)

Fix #20263