Qt: Parent the ProfileSwitcher and the PacketDiagram QGraphicsScene
Parent the ProfileSwitcher and the PacketDiagram's QGraphicsScene
appropriately when creating them to prevent leaks. (This doesn't
exactly leak currently according to some leak analyzers, because
we never exit by returning from main but instead call exit via
exit_application, which ends up never running the main application
destructor so it doesn't get reported as a leak because the memory
is still in use. See #20370 and #14395.)