Lua: Don't allow negative offsets to TvbRange_range
Use the same restrictions on offset and len in TvbRange_range
as already used in TvbRange_raw:

Don't allow negative offsets - previously, this would work if the
TvbRange was a subset of a parent backing Tvb, by escaping into
the parent offset. This was undocumented and not how the epan
functions behave).

Treat length -1 as "to the end of this TvbRange" and not "to the
end of the parent Tvb" which was also different if the range was
a subset of the parent.

Immediately fail on other negative lengths.

Distinguish in error messages between "offset is too large and
runs off the end" and "offset plus length is too large and runs
off the end"

Ping #20103, #15655