Drop invalid malloc attribute from wmem_tree_new*
Wireshark would fail its wmem_test test when built with GCC LTO in some
cases with a null pointer dereference with a check that got optimised out.

This turns out to be because wmem_tree_new and wmem_tree_new_autoreset
are (incorrectly) marked with the malloc attribute, which promises to GCC
that any memory returned by the functions is completely clean and holds
no references to existing objects.

Similar to 365a8212989ef0ef506969631617175a2464302f.

See also:
* https://bugs.debian.org/1015706
* https://bugs.gentoo.org/941890
* https://gcc.gnu.org/PR117315

Closes #18216