dfvm: Move DFVM_VALUE_STRING opcode into READ_TREE
There are hfinfo with the same abbreviation but different value
strings. That means that we cannot read the tree and get a
field value, and then convert it with a value string in a subsequent
instruction using only information known at filter compile time. The
appropriate value string to use depends on the header field info
actually found in the frame. After we convert to a FVALUE, the
particular hfinfo is lost; only the abbreviation typed in the filter
remains, which will match only one particular field.

Eliminate the DFVM_VALUE_STRING opcode and incorporate the
"we want the value string for this field" information (already
a property of the syntax tree field nodes) into the DVFM_READ_
instructions, at which point we know the actual hfinfo in tree
and can select the proper value string.

Disallow at the semantic check having vals() operate on a field
that has had the raw operator ("@") applied to it, which makes
the effective field type FT_BYTES. (We could check sttype_field_ftenum
instead, but this gives a more complete error message. We also could
just ignore the raw operator and apply on the actual value, but
that seems like a bad idea.)

This makes filters like gsm_sms.tp-mti == "SMS-SUBMIT" and
gsm_sms.tp-mti == "SMS-SUBMIT REPORT" work properly, as well
vals(gsm_sms.tp-mti).

Fix #20399