LTE-RRC: Call E.212 MCC-MNC dissector function
Call the E212 dissector function to dissect the MCC and MNC
for a PLMN-Identity.

3GPP TS 36.331 notes that the MCC is optional when it is not the
first occurrence of the PLMN-Identity with in the PLMN-IdentityList,
and that if absent, it takes the same value as the MCC of the
immediately preceding PLMN-Identity (to handle MVNOs, etc.)
Store the most recently seen MCC to handle this situation.