dfilter: Don't crash on vals for BASE_UNIT_STRING
Make sure we check for BASE_UNIT_STRING and don't treat it like
an ordinary value string, with illegal memory accesses. For now,
it's not supported (would it be that useful?)

This applies to both the vals() function and filters that attempt
to match a numeric field with a string.

Before:

$ ./run/dftest 'ipv6.hopopts.len_oct == "test"'
Filter:
 ipv6.hopopts.len_oct == "test"

=================================================================
==3011913==ERROR: AddressSanitizer: global-buffer-overflow on address 0x7f4a7a177438 at pc 0x7f4a664738e9 bp 0x7fff9ca69b50 sp 0x7fff9ca69b48
READ of size 8 at 0x7f4a7a177438 thread T0
    #0 0x7f4a664738e8 in mk_fvalue_from_hfinfo /home/johnthacker/wireshark/epan/dfilter/semcheck.c:556:16

After:

$ ./run/dftest 'ipv6.hopopts.len_oct == "test"'
Filter:
 ipv6.hopopts.len_oct == "test"

Error: Unsigned integer (16 bits) cannot be converted from a string ("test").
  ipv6.hopopts.len_oct == "test"
                          ^~~~~~