wiretap: return file type extensions in listed order
Because of the way `wtap_get_file_extension_type_extensions` works,
the returned list has the extensions in the reverse order from how they're
listed in `file_type_extensions_base`. This reverse order is then shown to the
user in the file open dialog.

This behavior is mildly surprising, so fix it by reversing the list back into
the original order. For consistency, also do this in other functions returning
extension lists.