Qt: Fix tap parameter dialog expand all nodes check
The intent in TapParameterDialog is to expand all the nodes if
the total number of nodes in the tree is less than some arbitrary
number, currently 100.

ui->statsTreeWidget->model()->rowCount() doesn't give the total number
of rows in the tree, only the number of children of the root node.

Iterate over the tree (breaking if we hit our maximum) and expand
the nodes if there are that many total nodes.

Ping #20038