stats_tree: Update documentation
Clarify how the stats_tree flags are used. Some are for the entire
tree, some can be set on nodes, and some are internal use only. All
of them have to not overlap with the generic tapping system flags for
some reason.

Update the README for some recent changes adding tap flags to the
per packet routine and changing the directory separator to being a
double forward slash.

ST_FLG_SORT_DESC had backwards comment text, but has always indicated
that descending order was to be used (at least in the tshark CLI
tap, and in various functions, and in how the preference was interprted,
even if the Qt GUI didn't respect it previously.)

Also, packet-ltp was using the sort ordering incorrectly, and probably
didn't intend for counts to be ascending, so remove it. (It wasn't
doing anything other than actually requesting the tap columns, which
it didn't need, and the default order of descending count was used.)

Fix #20038