ftypes: Allow 0 to convert to a IEEE-11073 [S]Float
If the literal value consists of entirely leading zeros
(e.g., "0") with no fractional part, then *i_char already points
to the null terminator after skipping the leading zeroes, which
means we need to skip the while loop and not iterate i_char past
the null terminator.

On a ASAN build, the current code will stop with a buffer overflow.
On most builds, the filter will unexpectedly fail, claiming that
"0" cannot be converted to a IEEE-11073 float.

Just make the do..while loop into a simple while loop. (We can
also remove a check if c is '\0' because it won't enter the loop
in that case.)