bacapp: Handle segmented messages with more than 256 segments
Even if segments are out-of-order, if there are no permanently
missing segments the way the SegmentACK Window Size works it is
fine to use the last contiguously received segment number rather
than storing the highest number in conversation data. If segments
are permanently missing from the capture defragmentation won't
complete anyway.

Fix #20410