tls: Don't export PDUs segments desegmented in a later frame
TLS, like TCP, relies on the application dissector to determine
if desegmentation is needed. Don't export bytes that the subdissector
didn't actually handle, but indicated needed to be dissected in a
future frame. This happens for every fragmented case on the first
pass (so, when exporting PDUs from one-pass tshark). It also happens
on any pass when TLS contained in a TCP segment has at least one complete
record but also some segments that need to be reassembled later.
(The TLS dissector does not split MSPs like the TCP dissector now does,
which might remove the second pass issue, but the one-pass issue would
remain.)

Analogous to #15686.

This produces correct exported PDU with the sample in #20264.