TVB: Check that not too many bits are requested to be returned
With specific bit extraction functions, each having the appropriate
return type, it would be an error if the number of bits requested is
larger than the return type could contain.
Either the requested number of bits comes from unsanitised input, or
an other programming error. Either way the dissector shall no do this.