TVB: make the last tvb_get function adhere to encoding enums
It has been a long time since the boolean endianness encoding flag was
replaced by a more expressive bit field. This has the ENC_LITTLE_ENDIAN
value to indicate such encoding, among many others.

tvb_get_guid() is the last of the TVB access functions that uses the
boolean interpretation of the encoding field. Its comment suggests the
related proto_tree_add_item function depends on this, but the code for
that function reveils the boolean interpretation of its encoding flag
is already made there and replaced with ENC_LITTLE_ENDIAN if needed.

Therefore both the comment can be dropped and the expression aligned
with the other TVB access functions.