tcp: Don't check retransmission/ooo if desegment_tcp did
In certain cases, desegment_tcp can deal with out of order or
retransmitted segments even when the "reassemble out-of-order
segments" preference is off. This is primarily when two segments
are out of order relative to each other, but both lie within the
sequence number range of a multisegment PDU begun by an earlier,
in order segment. (E.g., after an HTTP Response Header with a large
Content-Length, subsequent segments all part of the content can
be received out of order but still reassembled properly, as it is
already known that they belong to the same PDU.)

desegment_tcp has checks, both when out of order reassembly is on
and off, for when TCP analysis has set a retransmission or out of
order flag. In the case where the segment is known to be able to be
used in a multisegment PDU (and thus the final stream can still be
delivered in order), retransmitted segments are still dissected and
the flag effectively ignored.

The check currently in decode_tcp_ports is superfluous to the checks
already in desegment_tcp, except that it (when out of order reassembly
is off) incorrectly refuses to dissect in cases where the segment is
the final segment received of such an MSP and in-order delivery still
achieved.

Since desegment_tcp has superior checking for reassembled and out
of order segments in all cases, move that retransmission check to
process_tcp_payload, and perform it in the cases where desegment_tcp
was not called, along with other existing checks of whether a segment
is in the middle of a PDU.

Fix #20406, which has a case where this happens. (In the capture on
that issue, frame 205 has a sequence number earlier than that of frame
203, but is the last segment received to complete the HTTP response
begun in frame 11. Despite the out of order frames in the capture, the
TCP dissector can properly deliver the stream in order to the HTTP
dissector even with "reassemble out-of-order" disabled.)