sharkd: Update tests that use out-of-order capture
The http-ooo.pcap requires setting the tcp.reassemble_out_of_order
preference to true to dissect correctly. While there is sometimes
value in testing that capture with the preference off, consistent
results from release to release are not guaranteed (improvements to
how out of order dissection is attempted with the preference off can
change the results.) Out of order testing is done in the detection
test suite. The sharkd suite should set the preference on so as to
have consistent results that only test the sharkd behavior.