HTTP: Add preference for heuristic first (MIME sniffing)
RFC 9110 8.3 Content Type:
"In practice, resource owners do not always properly configure their origin
server to provide the correct Content-Type for a given representation. Some
user agents examine the content and, in certain cases, override the received
type... When sniffing is implemented, implementers are encouraged to provide
a means for the user to disable it."

Similarly RFC 7231 3.1.1.5 Content Type.

Cf. RFC 2616 (obsoleted) 7.2.1 Type:
"If and only if the media type is not given by a Content-Type field, the
recipient MAY attempt to guess the media type via inspection of its
content and/or the name extension(s) of the URI used to identify the
resource."

Also see https://mimesniff.spec.whatwg.org/

Since the Content-Type is not always correct (and in some cases the
there may be a reason to dissect a type like "application/octet-stream"
with a specific heuristic dissector), add a preference to try heuristic
dissectors before the Content-Type. Disable it by default, for
compliance with best practices as expressed in RFCs 7231 and 9110.

This should also perhaps do the ports first too. (See #17295, though
perhaps a well-craft heuristic would work there too.)