ITS: CPM Fix extensibility flag of WrappedCpmContainer
asn2wr script generates C code marking this list as extensible.
However, further constraining the field with WITH COMPONENT(S)
removes this bit. To fix this, the flag is explicitly removed.

As some ITS implementations asn1 generators create faulty code a
preference is introduce which allows for the previous dissection.