Bluetooth: Validate UUID format for the UAT
Validate the UUID format for custom UUID names, and provide an
error message. Make sure that the names are converted to lowercase
before putting in the tree, because that's what
print_numeric_bluetooth_uuid does.

The Bluetooth UUIDs should probably be a FT_GUID variant, compatible
but with special handling for translating the 16-bit UUIDs.

Fix #20419