package: Update RPM spec for optional dependencies
The rpm-setup.sh script in tools installs optional dependencies on
RPM-based distributions.

We enable optional dependencies in the RPM package if CMake finds them
(this requires having the feature enabled at least once, though it can
be cached if the same build directory is reused). The RPM spec has to
support each option with a bcond_ directive, have BuildRequires if
enabled (so that mechanisms like mock work correctly), and also disable
the feature if --without is given. Update that for a number of the
packages, and try to put them in a more consistent order across the
various places.

Some of these optional dependencies that have been present in all
distributions for a long time should probably default to on.

Ping #20422