packaging: Move COPYING file from doc to license in rpm
RPM spec files (since rpm 4.11, and we require 4.13) have a different
file attribute for license files instead of documentation. The
difference is that documentation files can be optionally omitted when
installing to save space, but license files cannot be filtered out.
Move the copy of the GPL version 2 to license.