Qt: Make TCPStreamDialog event-driven, prevent crash on closing
Make the TCP Stream Dialog react to the CaptureEvents when retapping
is done. This has the advantage that the signal isn't received when
the TCPStreamDialog has been destroyed, instead of always manipulating
the UI elements after the potentially very long retap is finished.
This uses the CaptureFile QT object instead of the capture_file pointer
directly.

The TCP Stream Dialog taps and then disconnects, instead of staying
connected as a tap. Continue with that behavior, but allow the existing
graph to still be changed after the file is closed (aside from the
stream number, which requires a re-tap.)

Disable the stream number spinbox during tapping with setReadOnly, which
achieves the goal of disabling it but not affecting focus more simply.
Completely disable it when the file is closed, because it won't be
enabled again.

Remove the tap listener when destroying the dialog if we're still tapping.
(depends on c79d79f2e16c9e6bd9a3d9a9837cfbff13300c4f) This will produce
a console warning when tapping finishes (no listener found with that tap
data) if the dialog has already been closed, but is better than crashing.
We could remove the tap listener upon the captureEvent in the Qt dialog
to avoid the console warning instead of in graph_segment_list_get,
changing its behavior.

Free the graph segments as the tap reset callback, which is a little
cleaner.

Fix #20426