tap-tcp: Remove internal struct
The tcp_scan_t struct is internal and doesn't seem to be helpful.
The "direction" int is initialized to a fixed value when the struct is
created and never changed. Move the pointer to the last segment in the
list (used for appending) to the externally facing tcp_graph struct.
(As noted, the list could possibly be a different structure.)

Remove it, and just use the externally facing tcp_graph as the tapinfo.
This is helpful for #20426 because it can allow the TCP Stream Dialog
to remove the tap listener, which is needs to do because it is the
parent of the tcp_graph struct. The dialog needs to remove the tap listener
before freeing the graph memory upon destruction (and to stop the
calculations), when the dialog is closed before tapping is finished.