Signal-PDU, Some/IP: Deregister fields and their arrays later
Don't call proto_free_deregistered_fields() from a dissector,
including in prefs, because there can be a tree which refers
to its dynamically registered fields (perhaps interesting fields
to a Lua field extractor.) Wireshark will free those fields at
an appropriate safe time, such as when redissecting.

We still need to set the dynamic hf arrays to NULL so that the
dissector doesn't refer to them, but don't want to leak them. That
is what proto_add_deregistered_data is for - it will free those
arrays at the time that the fields are actually freed.

Fix #20425