ELF: Add heuristics over HTTP
ELF has a four byte magic number, not all printable ASCII
(the first is a 0x7F/DEL), which should be sufficient for a
heuristic.

The capture file in #20406 is an example of an ELF file with the
wrong Content-Type. (The new "try heuristic subdissectors first"
preference for HTTP would also need to be enabled.)