ftypes: Restore ability to filter and compare negative IEEE-11073 floats
The changes before 4.0 that made the minus sign a separate token
ultimately made it impossible to filter negative IEEE-11073 floats,
because unary minus was not defined (all negative numbers are positive
numbers plus unary minus.)

In order to support negative numbers, we need to implement unary
minus, and the "is negative" test. Also fix the "is zero" test so
that it is correct for all the different values of zero.

Add a few tests for IEEE 11073 floats, adding a capture with two frames
with SFLOATs and FLOATs and the ability to specify Decode As for the
dfilter test suite in order to decode it without the other Bluetooth
frames.

Ping #19011